refactor(governance): break down into smaller files
