+++ packages/sdk-governance/src/index.ts
@@ -1,12 +1,12 @@
-export { createReferendaSdk } from "./referenda-sdk"
+export { createReferendaSdk } from "./referenda/referenda-sdk"
 export {
   type Origin,
   polkadotSpenderOrigin,
   kusamaSpenderOrigin,
-} from "./referenda-chainConfig"
-export type * from "./referenda-descriptors"
-export type * from "./referenda-sdk-types"
+} from "./referenda/chainConfig"
+export type * from "./referenda/descriptors"
+export type * from "./referenda/sdk-types"
 
-export { createBountiesSdk } from "./bounties-sdk"
-export type * from "./bounties-descriptors"
-export type * from "./bounties-sdk-types"
+export { createBountiesSdk } from "./bounties/bounties-sdk"
+export type * from "./bounties/descriptors"
+export type * from "./bounties/sdk-types"
